WOOC-716 prevent error if option is already deleted

+++ uninstall.php
@@ -7,7 +7,9 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 
 $option_name = 'woocommerce_payplug_settings';
 
-delete_option( $option_name );
+if( get_option('woocommerce_payplug_settings') ){
+	delete_option( $option_name );
+	// for site options in Multisite
+	delete_site_option( $option_name );
+}
 
-// for site options in Multisite
-delete_site_option( $option_name );
